+++ packages/types/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.8.0] - 2024-07-03
 ### Added
 - Add type `EthereumNetworkModule` to support eth module (#318)
 
@@ -121,7 +122,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sync with main sdk (#14)
 
 ## [0.1.0] - 2022-10-31
-[Unreleased]: https://github.com/subquery/subql-ethereum/compare/types-ethereum/3.7.0...HEAD
+[Unreleased]: https://github.com/subquery/subql-ethereum/compare/types-ethereum/3.8.0...HEAD
+[3.8.0]: https://github.com/subquery/subql-ethereum/compare/types-ethereum/3.7.0...types-ethereum/3.8.0
 [3.7.0]: https://github.com/subquery/subql-ethereum/compare/types-ethereum/3.6.1...types-ethereum/3.7.0
 [3.6.1]: https://github.com/subquery/subql-ethereum/compare/types-ethereum/3.6.0...types-ethereum/3.6.1
 [3.6.0]: https://github.com/subquery/subql-ethereum/compare/types-ethereum/3.5.0...types-ethereum/3.6.0